Move 2 future indent functions from editor to format. Pass manual tests.
